move VGGLayer to Application
