+++ src/Application/CMakeLists.txt
@@ -23,6 +23,7 @@ add_library(vgg_application STATIC
   VggEnv.cpp
   VggSdk.cpp
   ViewModel.cpp
+  VGGLayer.cpp
 )
 
 if(MSVC)
